+++ src/main/java/com/StockPulse/StockPulse/dao/UserDao.java
@@ -21,6 +21,6 @@ public interface UserDao {
 
 
     // TODO - Delete User
-    void deleteUser(int userId);
+    void deleteUser(Long userId);
 
 }
